refactor: split god files, reduce function length, fix DRY violations
- agent/mod.rs: extract pre_process_security, advance_context_lifecycle,
  build_user_message from process_user_message_inner (H2)
- openai/mod.rs: replace inline retry with shared send_with_retry (H5)
- subagent/manager.rs: extend create_transcript_writer with resumed_from,
  replace inline executor setup in resume() with build_filtered_executor (H3)
- router/mod.rs: extract cascade_evaluate_response helper to eliminate
  duplicate evaluate/record/log blocks in cascade_chat and cascade_chat_stream (MISSED-2)
- zeph-sanitizer: extract all trust model and output types to types.rs (M5)
- zeph-tui: extract RenderCache* to render_cache.rs, message types to types.rs (M1)
- learning.rs: fix private_interfaces clippy lint on InferredPreference (M2 partial)
